move ValidationNodeConfig to separate file
